+++ Runtime/AbstractAnimation.cs
@@ -34,6 +34,9 @@ namespace OneM.AnimationSystem
 
         private void OnDisable() => Stop();
 
+        public void ResetSpeed() => speed = 1F;
+        public void RevertSpeed() => speed *= -1F;
+
         public void Restart()
         {
             Stop();